+++ modules/core-config/modules/kube-prometheus-stack/identity.tf
@@ -27,3 +27,21 @@ module "identity_grafana" {
 
   tags = var.tags
 }
+
+module "identity_thanos" {
+  source = "../../../identity"
+
+  location            = var.location
+  resource_group_name = var.resource_group_name
+
+  name      = "${var.cluster_name}-thanos"
+  namespace = var.namespace
+  labels    = var.labels
+
+  roles = [{
+    id    = "Storage Blob Data Contributor"
+    scope = "${var.storage_account_id}"
+  }]
+
+  tags = var.tags
+}