+++ modules/core-config/modules/kube-prometheus-stack/local.tf
@@ -1,6 +1,11 @@
 locals {
   chart_version = "39.9.0"
 
+  thanos_chart_version = "1.5.0"
+
+  # Thanos image version should match version in Thanos chart
+  thanos_image_version = "0.27.0"
+
   chart_values = {
     global = {
       rbac = {
@@ -15,6 +20,10 @@ locals {
 
     prometheusOperator = {
       enabled = true
+
+      tls = {
+        internalPort = 10250
+      }
 
       priorityClassName = ""
 
@@ -57,15 +66,17 @@ locals {
           }
 
           limits = {
-            cpu    = "050m"
+            cpu    = "500m"
             memory = "16Mi"
           }
         }
       }
 
       admissionWebhooks = {
+        enabled = true
+
         patch = {
-          priorityClassName = ""
+          priorityClassName = "system-cluster-critical"
 
           nodeSelector = {
             "kubernetes.io/os" = "linux"
@@ -92,7 +103,7 @@ locals {
 
     prometheus = {
       prometheusSpec = {
-        retention = "28d"
+        retention = "6h"
 
         remoteWrite = var.prometheus_remote_write
 
@@ -118,11 +129,11 @@ locals {
         logLevel  = "info"
         logFormat = "json"
 
-        priorityClassName = ""
+        replicas = var.zones
 
-        podMetadata = {
-          labels = var.labels
-        }
+        replicaExternalLabelName = "prometheus_replica"
+
+        priorityClassName = "system-cluster-critical"
 
         nodeSelector = {
           "kubernetes.io/os" = "linux"
@@ -143,6 +154,15 @@ locals {
         podAntiAffinity            = "hard"
         podAntiAffinityTopologyKey = "topology.kubernetes.io/zone"
 
+        podMetadata = {
+          labels = merge(var.labels, {
+            aadpodidbinding = module.identity_thanos.name
+          })
+          annotations = {
+            "checksum/thanos-objstore-config" = local.thanos_objstore_secret_checksum
+          }
+        }
+
         storageSpec = {
           volumeClaimTemplate = {
             spec = {
@@ -154,7 +174,7 @@ locals {
 
               resources = {
                 requests = {
-                  storage = "512Gi"
+                  storage = "64Gi"
                 }
               }
             }
@@ -172,6 +192,45 @@ locals {
             memory = "4096Mi"
           }
         }
+
+        thanos = {
+          image = "quay.io/thanos/thanos:v${local.thanos_image_version}"
+
+          resources = {
+            requests = {
+              cpu    = "50m"
+              memory = "64Mi"
+            }
+
+            limits = {
+              cpu    = "1000m"
+              memory = "64Mi"
+            }
+          }
+
+          objectStorageConfig = {
+            name = local.thanos_objstore_secret_name
+            key  = local.thanos_objstore_secret_key
+          }
+        }
+      }
+
+      serviceAccount = {
+        create = true
+      }
+
+      thanosService = {
+        enabled = true
+      }
+
+      thanosServiceMonitor = {
+        enabled = true
+      }
+
+      podDisruptionBudget = {
+        enabled        = true
+        minAvailable   = ""
+        maxUnavailable = 1
       }
 
       ingress = {
@@ -317,7 +376,22 @@ locals {
         "grafana-piechart-panel"
       ], var.grafana_additional_plugins))
 
-      additionalDataSources = concat([local.grafana_azure_monitor_data_source], var.grafana_additional_data_sources)
+      additionalDataSources = concat([local.grafana_azure_monitor_data_source], [for datasource in var.grafana_additional_data_sources : {
+        name            = datasource.name
+        type            = datasource.type
+        access          = datasource.access
+        orgId           = lookup(datasource, "orgId", null)
+        uid             = lookup(datasource, "uid", null)
+        url             = lookup(datasource, "url", null)
+        basicAuth       = lookup(datasource, "basicAuth", null)
+        basicAuthUser   = lookup(datasource, "basicAuthUser", null)
+        withCredentials = lookup(datasource, "withCredentials", null)
+        isDefault       = lookup(datasource, "isDefault", null)
+        jsonData        = lookup(datasource, "jsonData", null)
+        secureJsonData  = lookup(datasource, "secureJsonData", null)
+        version         = lookup(datasource, "version", null)
+        editable        = lookup(datasource, "editable", null)
+      }])
 
       priorityClassName = ""
 
@@ -361,7 +435,18 @@ locals {
         }
       }
 
+      serviceMonitor = {
+        enabled = true
+        labels = {
+          "lnrs.io/monitoring-platform" = "true"
+        }
+      }
+
       sidecar = {
+        datasources = {
+          url = "http://thanos-query-frontend.${var.namespace}.svc.cluster.local:10902"
+        }
+
         dashboards = {
           enabled          = true
           folderAnnotation = "grafana_folder"
@@ -372,7 +457,6 @@ locals {
           }
           searchNamespace = "ALL"
         }
-
 
         resources = {
           requests = {
@@ -414,7 +498,6 @@ locals {
       create = false
     }
 
-    ## Deploy servicemonitor
     kubeStateMetrics = {
       enabled = true
     }
@@ -433,7 +516,7 @@ locals {
         }
       }
 
-      priorityClassName = ""
+      priorityClassName = "system-cluster-critical"
 
       customLabels = var.labels
 
@@ -519,6 +602,523 @@ locals {
     }
   }
 
+  thanos_chart_values = {
+    serviceMonitor = {
+      enabled = true
+      additionalLabels = {
+        "lnrs.io/monitoring-platform" = "true"
+      }
+    }
+
+    commonLabels = var.labels
+
+    objstoreConfig = {
+      create = false
+      name   = local.thanos_objstore_secret_name
+      key    = local.thanos_objstore_secret_key
+    }
+
+    logLevel  = "info"
+    logFormat = "json"
+
+    compact = {
+      enabled = true
+
+      serviceAccount = {
+        create = true
+      }
+
+      podLabels = {
+        aadpodidbinding = module.identity_thanos.name
+      }
+
+      podAnnotations = {
+        "checksum/thanos-objstore-config" = local.thanos_objstore_secret_checksum
+      }
+
+      priorityClassName = ""
+
+      resources = {
+        requests = {
+          cpu    = "100m"
+          memory = "512Mi"
+        }
+
+        limits = {
+          cpu    = "1000m"
+          memory = "512Mi"
+        }
+      }
+
+      persistence = {
+        enabled      = true
+        storageClass = "azure-disk-premium-ssd-delete"
+        accessMode   = "ReadWriteOnce"
+        size         = "64Gi"
+      }
+
+      nodeSelector = {
+        "kubernetes.io/os" = "linux"
+        "lnrs.io/tier"     = "system"
+      }
+
+      tolerations = [
+        {
+          key      = "CriticalAddonsOnly"
+          operator = "Exists"
+        },
+        {
+          key      = "system"
+          operator = "Exists"
+        }
+      ]
+
+      extraArgs = [
+        "--retention.resolution-raw=28d",
+        "--retention.resolution-1h=28d",
+        "--retention.resolution-5m=28d",
+        "--delete-delay=72h"
+      ]
+    }
+
+    query = {
+      serviceAccount = {
+        create = true
+      }
+
+      autoscaling = {
+        enabled                        = true
+        minReplicas                    = 1
+        maxReplicas                    = 3
+        targetCPUUtilizationPercentage = 80
+      }
+
+      podDisruptionBudget = {
+        enabled        = true
+        maxUnavailable = "33%"
+      }
+
+      updateStrategy = {
+        type = "RollingUpdate"
+
+        rollingUpdate = {
+          maxUnavailable = "33%"
+          maxSurge       = 0
+        }
+      }
+
+      priorityClassName = ""
+
+      replicaLabels = ["prometheus_replica"]
+      additionalStores = [
+        "dnssrv+_grpc._tcp.kube-prometheus-stack-thanos-discovery.${var.namespace}.svc.cluster.local"
+      ]
+
+      resources = {
+        requests = {
+          cpu    = "500m"
+          memory = "1024Mi"
+        }
+
+        limits = {
+          cpu    = "2000m"
+          memory = "1024Mi"
+        }
+      }
+
+      nodeSelector = {
+        "kubernetes.io/os" = "linux"
+        "lnrs.io/tier"     = "system"
+      }
+
+      tolerations = [
+        {
+          key      = "CriticalAddonsOnly"
+          operator = "Exists"
+        },
+        {
+          key      = "system"
+          operator = "Exists"
+        }
+      ]
+
+      affinity = {
+        podAntiAffinity = {
+          preferredDuringSchedulingIgnoredDuringExecution = [
+            {
+              podAffinityTerm = {
+                labelSelector = {
+                  matchLabels = {
+                    "app.kubernetes.io/name"      = "thanos"
+                    "app.kubernetes.io/instance"  = "thanos"
+                    "app.kubernetes.io/component" = "query"
+                  }
+                }
+
+                topologyKey = "topology.kubernetes.io/zone"
+              }
+
+              weight = 100
+            }
+          ]
+        }
+      }
+
+      extraArgs = [
+        "--query.timeout=5m",
+        "--query.lookback-delta=15m"
+      ]
+    }
+
+    queryFrontend = {
+      enabled = true
+
+      serviceAccount = {
+        create = true
+      }
+
+      autoscaling = {
+        enabled                        = true
+        minReplicas                    = 1
+        maxReplicas                    = 3
+        targetCPUUtilizationPercentage = 80
+      }
+
+      podDisruptionBudget = {
+        enabled        = true
+        maxUnavailable = "33%"
+      }
+
+      updateStrategy = {
+        type = "RollingUpdate"
+
+        rollingUpdate = {
+          maxUnavailable = "33%"
+          maxSurge       = 0
+        }
+      }
+
+      priorityClassName = ""
+
+      ingress = {
+        enabled          = true
+        annotations      = var.ingress_annotations
+        ingressClassName = var.ingress_class_name
+        pathType         = "Prefix"
+        hosts            = ["thanos-${var.ingress_subdomain_suffix}.${var.ingress_domain}"]
+      }
+
+      resources = {
+        requests = {
+          cpu    = "100m"
+          memory = "64Mi"
+        }
+
+        limits = {
+          cpu    = "1000m"
+          memory = "64Mi"
+        }
+      }
+
+      nodeSelector = {
+        "kubernetes.io/os" = "linux"
+        "lnrs.io/tier"     = "system"
+      }
+
+      tolerations = [
+        {
+          key      = "CriticalAddonsOnly"
+          operator = "Exists"
+        },
+        {
+          key      = "system"
+          operator = "Exists"
+        }
+      ]
+
+      affinity = {
+        podAntiAffinity = {
+          preferredDuringSchedulingIgnoredDuringExecution = [
+            {
+              podAffinityTerm = {
+                labelSelector = {
+                  matchLabels = {
+                    "app.kubernetes.io/name"      = "thanos"
+                    "app.kubernetes.io/instance"  = "thanos"
+                    "app.kubernetes.io/component" = "query-frontend"
+                  }
+                }
+
+                topologyKey = "topology.kubernetes.io/zone"
+              }
+
+              weight = 100
+            }
+          ]
+        }
+      }
+
+      extraArgs = [
+        "--labels.split-interval=12h",
+        "--labels.max-retries-per-request=10",
+        "--query-range.split-interval=12h",
+        "--query-range.align-range-with-step",
+        "--query-range.max-retries-per-request=10",
+        "--query-frontend.compress-responses",
+        "--query-frontend.log-queries-longer-than=10s",
+        "--query-frontend.downstream-tripper-config=${local.thanos_query_frontend_downstream_tripper_config}"
+      ]
+    }
+
+    rule = {
+      enabled = true
+
+      serviceAccount = {
+        create = true
+      }
+
+      podLabels = {
+        aadpodidbinding = module.identity_thanos.name
+      }
+
+      podAnnotations = {
+        "checksum/thanos-objstore-config" = local.thanos_objstore_secret_checksum
+      }
+
+      priorityClassName = ""
+
+      ingress = {
+        enabled          = true
+        annotations      = var.ingress_annotations
+        ingressClassName = var.ingress_class_name
+        pathType         = "Prefix"
+        hosts            = ["thanos-rule-${var.ingress_subdomain_suffix}.${var.ingress_domain}"]
+      }
+
+      resources = {
+        requests = {
+          cpu    = "100m"
+          memory = "128Mi"
+        }
+
+        limits = {
+          cpu    = "1000m"
+          memory = "128Mi"
+        }
+      }
+
+      persistence = {
+        enabled      = true
+        storageClass = "azure-disk-premium-ssd-delete"
+        accessMode   = "ReadWriteOnce"
+        size         = "16Gi"
+      }
+
+      nodeSelector = {
+        "kubernetes.io/os" = "linux"
+        "lnrs.io/tier"     = "system"
+      }
+
+      tolerations = [
+        {
+          key      = "CriticalAddonsOnly"
+          operator = "Exists"
+        },
+        {
+          key      = "system"
+          operator = "Exists"
+        }
+      ]
+
+      configReloader = {
+        enabled = true
+
+        resources = {
+          requests = {
+            cpu    = "10m"
+            memory = "16Mi"
+          }
+
+          limits = {
+            cpu    = "500m"
+            memory = "16Mi"
+          }
+        }
+      }
+
+      alertmanagersConfig = {
+        create = true
+
+        value = <<-EOT
+          alertmanagers:
+            - static_configs:
+                - kube-prometheus-stack-alertmanager.${var.namespace}.svc.cluster.local:9093
+              scheme: http
+         EOT
+      }
+
+      rules = {
+        create = false
+        name   = "thanos-ruler-thanos-ruler-rulefiles-0"
+      }
+
+      evalInterval = "5m"
+    }
+
+    storeGateway = {
+      serviceAccount = {
+        create = true
+      }
+
+      podLabels = {
+        aadpodidbinding = module.identity_thanos.name
+      }
+
+      podAnnotations = {
+        "checksum/thanos-objstore-config" = local.thanos_objstore_secret_checksum
+      }
+
+      replicas = var.zones
+
+      podDisruptionBudget = {
+        enabled        = true
+        maxUnavailable = 1
+      }
+
+      priorityClassName = ""
+
+      resources = {
+        requests = {
+          cpu    = "500m"
+          memory = "2048Mi"
+        }
+
+        limits = {
+          cpu    = "1000m"
+          memory = "2048Mi"
+        }
+      }
+
+      persistence = {
+        enabled      = true
+        storageClass = "azure-disk-premium-ssd-delete"
+        accessMode   = "ReadWriteOnce"
+        size         = "16Gi"
+      }
+
+      nodeSelector = {
+        "kubernetes.io/os" = "linux"
+        "lnrs.io/tier"     = "system"
+      }
+
+      tolerations = [
+        {
+          key      = "CriticalAddonsOnly"
+          operator = "Exists"
+        },
+        {
+          key      = "system"
+          operator = "Exists"
+        }
+      ]
+
+      affinity = {
+        podAntiAffinity = {
+          requiredDuringSchedulingIgnoredDuringExecution = [{
+            labelSelector = {
+              matchLabels = {
+                "app.kubernetes.io/name"      = "thanos"
+                "app.kubernetes.io/instance"  = "thanos"
+                "app.kubernetes.io/component" = "store-gateway"
+              }
+            }
+            topologyKey = "topology.kubernetes.io/zone"
+          }]
+        }
+      }
+    }
+  }
+
+  thanos_ruler = {
+    apiVersion = "monitoring.coreos.com/v1"
+    kind       = "ThanosRuler"
+
+    metadata = {
+      name      = "thanos-ruler"
+      namespace = var.namespace
+      labels    = var.labels
+    }
+
+    spec = {
+      replicas = 0
+
+      priorityClassName = "system-cluster-critical"
+
+      nodeSelector = {
+        "kubernetes.io/os" = "linux"
+        "lnrs.io/tier"     = "system"
+      }
+
+      tolerations = [
+        {
+          key      = "system"
+          operator = "Exists"
+        }
+      ]
+
+      storage = {
+        volumeClaimTemplate = {
+          spec = {
+            storageClassName = "ebs-gp3-retain"
+
+            accessModes : [
+              "ReadWriteOnce"
+            ]
+
+            resources = {
+              requests = {
+                storage = "10Gi"
+              }
+            }
+          }
+        }
+      }
+
+      resources = {
+        requests = {
+          cpu    = "100m"
+          memory = "128Mi"
+        }
+
+        limits = {
+          cpu    = "1000m"
+          memory = "128Mi"
+        }
+      }
+
+      logLevel  = "info"
+      logFormat = "json"
+
+      objectStorageConfig = {
+        name = local.thanos_objstore_secret_name
+        key  = local.thanos_objstore_secret_key
+      }
+
+      queryEndpoints = ["dnssrv+_http._tcp.thanos-query-frontend.${var.namespace}.svc.cluster.local"]
+
+      alertmanagersUrl = ["http://kube-prometheus-stack-alertmanager.${var.namespace}.svc.cluster.local:9093"]
+
+      ruleNamespaceSelector = {}
+
+      ruleSelector = {
+        matchLabels = {
+          "lnrs.io/monitoring-platform" = "true"
+          "lnrs.io/thanos-rule"         = "true"
+        }
+      }
+    }
+  }
+
   alertmanager_base_receivers = [{
     name              = "null"
     email_configs     = []
@@ -590,14 +1190,6 @@ locals {
     mute_time_intervals = lookup(route, "mute_time_intervals", [])
   }])
 
-  loki_data_source = {
-    name   = "Loki"
-    type   = "loki"
-    url    = "http://loki-gateway.logging.svc.cluster.local"
-    access = "proxy"
-    orgId  = "1"
-  }
-
   grafana_azure_monitor_data_source = {
     name      = "Azure Monitor"
     type      = "grafana-azure-monitor-datasource"
@@ -612,7 +1204,25 @@ locals {
   control_plane_log_analytics_workspace_external_resource_group_id = var.control_plane_log_analytics_workspace_different_resource_group ? regex("([[:ascii:]]*)(/providers/)", var.control_plane_log_analytics_workspace_id)[0] : ""
   oms_agent_log_analytics_workspace_resource_group_id              = var.oms_agent && var.oms_agent_log_analytics_workspace_id != null ? regex("([[:ascii:]]*)(/providers/)", var.oms_agent_log_analytics_workspace_id)[0] : ""
 
+  thanos_objstore_secret_name     = "thanos-object-storage"
+  thanos_objstore_secret_key      = "config"
+  thanos_objstore_secret_checksum = sha256(local.thanos_objstore_config)
+
+  thanos_objstore_config = <<-EOT
+    type: AZURE
+    config:
+      storage_account: ${var.storage_account_name}
+      container: ${azurerm_storage_container.thanos.name}
+      endpoint: blob.core.windows.net
+      user_assigned_id: ${module.identity_thanos.client_id}
+  EOT
+
+  thanos_query_frontend_downstream_tripper_config = <<-EOT
+    max_idle_conns_per_host: 100
+  EOT
+
   crd_files           = { for x in fileset(path.module, "crds/*.yaml") : basename(x) => "${path.module}/${x}" }
   resource_files      = { for x in fileset(path.module, "resources/*.yaml") : basename(x) => "${path.module}/${x}" }
+  resource_objects    = { thanos_ruler = local.thanos_ruler }
   dashboard_templates = { for x in fileset(path.module, "resources/configmap-dashboard-*.yaml.tpl") : basename(x) => { path = "${path.module}/${x}", vars = { resource_id = var.control_plane_log_analytics_workspace_id, subscription_id = var.subscription_id } } }
 }