+++ modules/core-config/modules/kube-prometheus-stack/main.tf
@@ -9,6 +9,19 @@ resource "kubectl_manifest" "crds" {
   wait              = true
 }
 
+resource "kubernetes_secret" "thanos_objstore_config" {
+  metadata {
+    name      = local.thanos_objstore_secret_name
+    namespace = var.namespace
+  }
+
+  type = "Opaque"
+
+  data = {
+    "${local.thanos_objstore_secret_key}" = local.thanos_objstore_config
+  }
+}
+
 resource "kubernetes_secret" "grafana_auth" {
   metadata {
     name      = "grafana-auth"
@@ -17,6 +30,7 @@ resource "kubernetes_secret" "grafana_auth" {
   }
 
   type = "Opaque"
+
   data = {
     "admin-user"     = "admin"
     "admin-password" = var.grafana_admin_password
@@ -27,6 +41,19 @@ resource "kubectl_manifest" "resource_files" {
   for_each = local.resource_files
 
   yaml_body = file(each.value)
+
+  server_side_apply = true
+  wait              = true
+
+  depends_on = [
+    kubectl_manifest.crds
+  ]
+}
+
+resource "kubectl_manifest" "resource_objects" {
+  for_each = local.resource_objects
+
+  yaml_body = yamlencode(each.value)
 
   server_side_apply = true
   wait              = true
@@ -66,6 +93,35 @@ resource "helm_release" "default" {
   ]
 
   depends_on = [
-    kubectl_manifest.crds
+    kubectl_manifest.crds,
+    azurerm_storage_container.thanos,
+    module.identity_thanos,
+    kubernetes_secret.thanos_objstore_config,
+    kubernetes_secret.grafana_auth
+  ]
+}
+
+resource "helm_release" "thanos" {
+  name      = "thanos"
+  namespace = var.namespace
+
+  repository = "https://stevehipwell.github.io/helm-charts/"
+  chart      = "thanos"
+  version    = local.thanos_chart_version
+  skip_crds  = true
+
+  max_history = 10
+  timeout     = 600
+
+  values = [
+    yamlencode(local.thanos_chart_values)
+  ]
+
+  depends_on = [
+    kubectl_manifest.crds,
+    helm_release.default,
+    azurerm_storage_container.thanos,
+    module.identity_thanos,
+    kubernetes_secret.thanos_objstore_config
   ]
 }