+++ modules/core-config/variables.tf
@@ -98,6 +98,11 @@ variable "oms_agent_create_configmap" {
   type        = bool
 }
 
+variable "storage_account_id" {
+  description = "ID of the data storage account."
+  type        = string
+}
+
 variable "labels" {
   description = "Labels to be applied to all Kubernetes resources."
   type        = map(string)